refactor: IOView InputView, OutputView로 리펙토링
