+++ src/main/java/Application.java
@@ -4,6 +4,8 @@ import domain.service.CardDistributor;
 import domain.service.ProfitCalculator;
 import domain.service.RandomCardGenerator;
 import view.IOView;
+import view.InputView;
+import view.OutputView;
 
 public class Application {
 
@@ -12,7 +14,11 @@ public class Application {
     }
 
     private static BlackJackController makeBlackJackController() {
-        return new BlackJackController(makeBlackJackGame(), new IOView());
+        return new BlackJackController(makeBlackJackGame(), makeIOView());
+    }
+
+    private static IOView makeIOView() {
+        return new IOView(new InputView(), new OutputView());
     }
 
     private static BlackJackGame makeBlackJackGame() {
